chore(events)!: split nats + kafka backends into separate crates
Two new workspace crates own the cloud-events publisher backends; the
core `lakekeeper` crate keeps only the `CloudEventBackend` trait, the
in-process `TracingPublisher`, and the audit backend. Same pattern as
the storage-postgres split (#1812).

### New crates

- `lakekeeper-events-nats`
  - `NatsBackend` + `build_nats_publisher_from_config()`
  - Owns `async-nats` and the 6 `nats_*` config fields
  - Backend-specific `config::CONFIG` static reads from the same
    `LAKEKEEPER__` / `ICEBERG_REST__` env prefixes — no operator-side
    changes required.

- `lakekeeper-events-kafka`
  - `KafkaBackend` + `KafkaConfig` + `build_kafka_publisher_from_config()`
  - Owns `rdkafka`, the 2 `kafka_*` config fields, and the previously
    in-core vendored `cloudevents-binding-rdkafka` (moved verbatim,
    history preserved via `git mv`)
  - Owns the `figment_file_provider_adapter` wiring for
    `LAKEKEEPER__KAFKA_CONFIG_FILE` (the only file-key adapter user)
  - `test_kafka_config_env_var` + `test_kafka_config_file` relocated
    from `lakekeeper/src/config.rs` to live with the type.

### `lakekeeper` cleanup

- Features removed: `nats`, `kafka`, `vendored-protoc` (empty feature,
  no in-tree consumer)
- Deps removed: `async-nats`, `rdkafka`, `figment_file_provider_adapter`
- Removed: `service/events/backends/{nats,kafka}/`, `pub use rdkafka`
- Removed 8 config fields (6 nats_*, 2 kafka_*) from `DynAppConfig`;
  dropped the now-empty `veil::Redact` derive (no fields left to redact)
- Removed `get_default_cloud_event_backends_from_config()` from
  `service::events::publisher`; replaced with a tiny
  `maybe_tracing_cloud_event_backend()` helper that returns the
  in-process tracing sink when `log_cloudevents` is on.

### `lakekeeper-bin`

- New `events.rs` module: binary-local
  `get_default_cloud_event_backends_from_config()` aggregates the three
  sources (nats crate, kafka crate, in-core tracing). Same pattern as
  `secrets.rs` / `get_default_catalog_from_config`.
- `serve.rs` imports the new helper from `crate::events::*`.
- `Cargo.toml`: drop the `nats,kafka,vendored-protoc` lakekeeper
  feature flags; add the two new path deps.

### Why

- Compile time: `rdkafka` (slow C-FFI dep) and `async-nats` no longer
  block every `cargo build` in the workspace. Toggling either on no
  longer forces a full `lakekeeper` rebuild. Default builds skip both.
- Clarity: identical pattern to storage-postgres — backend trait in
  core, concrete impl + config in named crate, binary wires them in.
- Removes the cross-module
  `pub kafka_config: Option<crate::service::events::backends::kafka::KafkaConfig>`
  leak from core config.
- The vendored `cloudevents-binding-rdkafka` is now isolated to the
  kafka crate — the "upstream PR or keep vendored" decision no longer
  pollutes the main tree.
