+++ crates/lakekeeper-bin/Cargo.toml
@@ -31,15 +31,11 @@ async-trait = { workspace = true }
 clap = { version = "^4.5", features = ["derive"] }
 figment = { workspace = true }
 http = { workspace = true }
-lakekeeper = { path = "../lakekeeper", features = [
-    "s3-signer",
-    "router",
-    "nats",
-    "vendored-protoc",
-    "kafka",
-] }
+lakekeeper = { path = "../lakekeeper", features = ["s3-signer", "router"] }
 lakekeeper-authz-openfga = { path = "../authz-openfga" }
 lakekeeper-console = { git = "https://github.com/lakekeeper/console", rev = "v0.14.3", optional = true }
+lakekeeper-events-kafka = { path = "../lakekeeper-events-kafka" }
+lakekeeper-events-nats = { path = "../lakekeeper-events-nats" }
 lakekeeper-secrets-kv2 = { path = "../lakekeeper-secrets-kv2" }
 lakekeeper-storage-postgres = { path = "../lakekeeper-storage-postgres" }
 limes = { workspace = true }
