+++ crates/lakekeeper-events-kafka/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "lakekeeper-events-kafka"
+version = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+description = "Kafka cloud-events publisher for Lakekeeper"
+keywords = ["iceberg", "rest", "lakekeeper", "kafka", "cloudevents"]
+
+[lib]
+
+[dev-dependencies]
+figment = { workspace = true, features = ["test"] }
+tempfile = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+async-trait = { workspace = true }
+cloudevents-sdk = { workspace = true }
+figment = { workspace = true }
+figment_file_provider_adapter = { workspace = true }
+lakekeeper = { path = "../lakekeeper", default-features = false }
+rdkafka = { workspace = true }
+serde = { workspace = true }
+tracing = { workspace = true }
+veil = { workspace = true }
