+++ crates/lakekeeper-events-nats/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "lakekeeper-events-nats"
+version = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+description = "NATS cloud-events publisher for Lakekeeper"
+keywords = ["iceberg", "rest", "lakekeeper", "nats", "cloudevents"]
+
+[lib]
+
+[dependencies]
+anyhow = { workspace = true }
+async-nats = { workspace = true }
+async-trait = { workspace = true }
+cloudevents-sdk = { workspace = true }
+figment = { workspace = true }
+lakekeeper = { path = "../lakekeeper", default-features = false }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tracing = { workspace = true }
+url = { workspace = true }
+veil = { workspace = true }
