+++ crates/lakekeeper-events-nats/src/lib.rs
@@ -1,13 +1,23 @@
+//! NATS cloud-events publisher for Lakekeeper.
+//!
+//! Implements [`lakekeeper::service::events::CloudEventBackend`] for a NATS
+//! client. Configured via env vars under the same `LAKEKEEPER__` /
+//! `ICEBERG_REST__` prefix as core; the [`config::CONFIG`] static aggregates
+//! the NATS-specific fields.
+
+pub mod config;
+
 use async_trait::async_trait;
 use cloudevents::Event;
+use lakekeeper::service::events::CloudEventBackend;
 
-use crate::{CONFIG, service::events::CloudEventBackend};
+use crate::config::CONFIG;
 
-/// Generate a NATS publisher from the crates configuration.
+/// Generate a NATS publisher from the crate's configuration.
 /// Returns `None` if the NATS address or topic is not set.
 ///
 /// # Errors
-/// - If NATs is configured but the connection or authentication fails.
+/// - If NATS is configured but the connection or authentication fails.
 pub async fn build_nats_publisher_from_config() -> anyhow::Result<Option<NatsBackend>> {
     let (Some(nats_addr), Some(nats_topic)) =
         (CONFIG.nats_address.clone(), CONFIG.nats_topic.clone())
