+++ crates/lakekeeper/src/service/events/backends/mod.rs
@@ -1,5 +1,1 @@
 pub mod audit;
-#[cfg(feature = "kafka")]
-pub mod kafka;
-#[cfg(feature = "nats")]
-pub mod nats;
